// src/frontend/src/api/model/baseWSProtocol.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BaseWSProtocolType } from './baseWSProtocolType';


export interface BaseWSProtocol { 
    baseProtocolType?: BaseWSProtocolType;
    /**
     * (don\'t care) : for internal protocol purpose
     */
    baseProtocolMsg?: string | null;
}
export namespace BaseWSProtocol {
}