// src/frontend/src/api/model/exampleWSProto2.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExampleWSProtocolType } from './exampleWSProtocolType';
import { BaseWSProtocolType } from './baseWSProtocolType';


export interface ExampleWSProto2 { 
    baseProtocolType?: BaseWSProtocolType;
    /**
     * (don\'t care) : for internal protocol purpose
     */
    baseProtocolMsg?: string | null;
    protocolType?: ExampleWSProtocolType;
    someLongValue?: number | null;
}
export namespace ExampleWSProto2 {
}