// src/frontend/src/api/model/exampleWSProtocol.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExampleWSProtocolType } from './exampleWSProtocolType';
import { BaseWSProtocolType } from './baseWSProtocolType';


export interface ExampleWSProtocol { 
    baseProtocolType?: BaseWSProtocolType;
    /**
     * (don\'t care) : for internal protocol purpose
     */
    baseProtocolMsg?: string | null;
    protocolType?: ExampleWSProtocolType;
}
export namespace ExampleWSProtocol {
}