// src/frontend/src/api/model/exampleWSProtoServerMem.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExampleWSProtocolType } from './exampleWSProtocolType';
import { BaseWSProtocolType } from './baseWSProtocolType';


export interface ExampleWSProtoServerMem { 
    baseProtocolType?: BaseWSProtocolType;
    /**
     * (don\'t care) : for internal protocol purpose
     */
    baseProtocolMsg?: string | null;
    protocolType?: ExampleWSProtocolType;
    /**
     * srv memory used in bytes
     */
    memoryUsed?: number | null;
}
export namespace ExampleWSProtoServerMem {
}